Local refactor
